Update dependency version com.fasterxml.jackson.core from version 2.9.9 to [2.9.9.2,)
Update pom.xml --> fix  vulnerable in com.fasterxml.jackson.core version 2.9.9, this dependency is upgraded to [2.9.9.2,)

A Polymorphic Typing issue was discovered in FasterXML jackson-databind 2.x before 2.9.9.2. This occurs when Default Typing is enabled (either globally or for a specific property) for an externally exposed JSON endpoint and the service has the logback jar in the classpath.